Add files via upload
In this project I have made an ML model which recognizes whether a parabola(with a lot of noise) is facing upward or downwards. This project is aim to test whether a NN is able to recognize  shapes and patterns from graph points. If our model succeeds we can use neural networks to find patterns in a stock market chart. Patterns in stock market charts help traders to predict the stock market.